+++ src/components/Profile.tsx
@@ -1,70 +1,47 @@
 import React, { FC, useState } from 'react';
-import { VStack, Box, Heading, Collapse, Button, Img } from '@chakra-ui/react';
+import {
+  VStack,
+  Box,
+  Heading,
+  Collapse,
+  Button,
+  Img,
+  Wrap,
+} from '@chakra-ui/react';
+import type { User } from '@/server/model/user';
 import ProfileField from './ProfileField';
+import ProfileBadge from './ProfileBadge';
 
 export interface ProfileProps {
-  avatarUrl: string;
-  username: string;
-  id: number;
-  nodeId: string;
-  gravatarId: string;
-  type: string;
-  siteAdmin: boolean;
-  name: string;
-  company: string;
-  blog: string;
-  location: string;
-  email: string;
-  hireable: boolean;
-  bio: string;
-  twitterUsername: string;
-  repositoryCount: number;
-  gistCount: number;
-  followerCount: number;
-  followingCount: number;
-  createdAt: string;
-  updatedAt: string;
+  shown: ShownProps;
+  hidden: User['minor'];
+  badges: User['badges'];
 }
 
-const HIDDEN_KEYS: (keyof ProfileProps)[] = [
-  'id',
-  'gravatarId',
-  'nodeId',
-  'type',
-  'siteAdmin',
-  'createdAt',
-  'updatedAt',
-];
+export interface ShownProps {
+  avatarUrl: string;
+  name: string;
+  login: string;
+  bio: string;
+  email: string;
+  location: string;
+  company: string;
+  websiteUrl: string;
+  twitterUsername: string;
+  repositories: number;
+  followers: number;
+  following: number;
+  gists: number;
+}
 
-const convertBoolean = (value: string | number | boolean) => {
-  if (typeof value === 'boolean') {
-    return value ? 'Yes' : 'No';
-  }
-
-  return value;
-};
-
-const Profile: FC<ProfileProps> = (props) => {
+const Profile: FC<ProfileProps> = ({ shown, hidden, badges }) => {
   const [show, setShow] = useState(false);
 
   const handleToggle = () => setShow((prev) => !prev);
-  const {
-    avatarUrl,
-    name,
-    children,
-    // hide the following,
-    id,
-    nodeId,
-    gravatarId,
-    type,
-    siteAdmin,
-    createdAt,
-    updatedAt,
-    ...rest
-  } = props;
+  const { avatarUrl, name, ...rest } = shown;
 
   return (
-    <VStack w={['90vw', '24rem']}>
+    <VStack w={['full', '24rem']}>
       <Img
         borderRadius="full"
         boxSize="200"
@@ -76,22 +53,19 @@ const Profile: FC<ProfileProps> = (props) => {
       <Heading as="h1" fontWeight="medium">
         {name}
       </Heading>
+      <Wrap>
+        {Object.keys(badges).map((key: keyof ProfileProps['badges']) => (
+          <ProfileBadge key={key} name={key} shown={badges[key]} />
+        ))}
+      </Wrap>
 
       <Box align="start" w="full">
-        {Object.keys(rest).map((key: keyof ProfileProps) => (
-          <ProfileField
-            key={key}
-            fieldKey={key}
-            fieldValue={convertBoolean(rest[key])}
-          />
+        {Object.keys(rest).map((key: keyof ShownProps) => (
+          <ProfileField key={key} fieldKey={key} fieldValue={rest[key]} />
         ))}
         <Collapse in={show}>
-          {HIDDEN_KEYS.map((key) => (
-            <ProfileField
-              key={key}
-              fieldKey={key}
-              fieldValue={convertBoolean(props[key])}
-            />
+          {Object.keys(hidden).map((key: keyof ProfileProps['hidden']) => (
+            <ProfileField key={key} fieldKey={key} fieldValue={hidden[key]} />
           ))}
         </Collapse>
       </Box>