+++ src/redux/api.ts
@@ -1,5 +1,4 @@
-import type { SearchUserPage } from '@/server/search-user';
-import type { User } from '@/server/user';
+import type { User, SearchUserPage } from '@/server/model/user';
 
 export interface SearchUserResponse
   extends LocalAPIResponse,